Add test_doesUserFollowAnotherUser() and rearrange : move test_articles() test_favouriteArticles() into UserTest
